Refactor battle resolution logic in MemedBattle and MemedBattleResolver contracts. Introduce checks for warrior NFT deployment and streamline score calculation methods. Update package.json deployment scripts for improved clarity. Enhance authorization checks in MemedFactory and MemedFactory_test contracts to include battle resolver. Remove unused NFT allocation function in MemedWarriorNFT for cleaner code.

+++ interfaces/IMemedBattleResolver.sol
@@ -4,6 +4,6 @@ pragma solidity ^0.8.28;
 import "../structs/BattleStructs.sol";
 
 interface IMemedBattleResolver {
-    function addBattleIdsToResolve(uint128 _battleId) external;
-    function resolveBattle(uint128 _battleId) external;
+    function addBattleIdsToResolve(uint256 _battleId) external;
+    function resolveBattle(uint256 _battleId) external;
 }
